Complete agent content verification contract
